Disable explicit Renovate PRs for boto3/botocore
We don't need to get explicit PRs for boto3/botocore, since we only
specify those packages as top-level dependencies to help Poetry's
dependency resolver.

We add the packages to `ignoreDeps` in `renovate.json`. With this
config, the two packages should still get updated once a week in the
lockfile maintenance.