Propagate exceptions raised while fetching dependencies
We need to collect the results from the iterator which is produced by
`ThreadPool.map()` to ensure that exceptions raised in the worker
threads are propagated to the main thread.

This commit collects the iterator by calling `list()` on it, which
ensures that any exceptions raised in a worker thread are propagated to
the main thread.

We also add a test case which verifies that `fetch_components()`
properly raises exceptions (e.g. when a requested component version
doesn't exist).