Use multiprocessing start method `spawn` by default
This is already the default on macOS, and doesn't break anything on
Linux. We'll need this at the latest when we want to switch to
gojsonnet, since that library doesn't work correctly with
multiprocessing start method `fork`.

Additionally, we reduce the memory usage of Commodore by using start
method `spawn` because the multiprocessing child processes are created
with smaller initial memory working sets. However, this causes some
execution time overhead because more data needs to be copied to the new
processes. Locally, we measured an increase in runtime of approximately
4 seconds when using `spawn` compared to `fork`.