Update `component new` and `component update` to expose the new template options
Add flags for each of the regex lists that allow configuring patch and
minor automerging behavior for the specified dependency names and
regex patterns.

For `component new` we add the set of flags that allow adding
dependency names and patterns to the lists, while for `component update`
we add that set of flags and an additional set of flags that allow
removing dependency names and patterns from the lists.