Clean up entry points provided by the Python package
We remove the legacy development commands `local_reveal` and `compile`.
These commands have never worked correctly when the package is installed
from PyPI since the package never bundled the `tools/tools.py` file.

Instead, we add a command `kapitan` which exposes the Kapitan CLI as an
entrypoint for the syn-commodore package. This command will enable users
who install Commodore with `uv tool install` to use the bundled Kapitan
as if it was an installed tool.