Explicitly handle redirects when doing POST requests
We Explicitly redo `POST` requests if we get a 301, 302, 307 or 308
status code for the first call. We don't validate that the redirect
location has the same domain as the original request, since we already
unconditionally follow redirects  with the bearer token for GET
requests.

Note that this wouldn't be necessary if all Lieutenant APIs would
redirect us with a 308 for POST requests.