Fix component multi-version verification
We need to check whether a component instance that sets a custom version
supports per-instance versions by checking the instance's target's
parameters. The bootstrap target ("cluster") doesn't have any
instance-specific parameters, but since we do the validation after
setting up all instance targets, we can use the instance's target to
read `_metadata` of the alias's version to properly validate that both
the base component and the alias are multi-version aware.

Additionally, because `component compile` doesn't have the regular
bootstrap target (and generating it just for alias validating seems a
bit overkill), we make the `bootstrap_target` key an optional parameter
for `verify_component_aliases()` and set it to the component that's
being compiled for `component compile` and fall back to
`config.inventory.boostrap_target` when it's not set.