Update tests
Adjust all tests that cover `verify_component_aliases()` to pass a
Kapitan/reclass inventory dict instead of just a parameters dict.

Additionally, update test inputs to set the alias `_metadata` in the
component parameters key instead of the instance parameters key to
replicate the structure that we generate for `catalog compile`.