Parallelize alias checkout and setup
We collect aliases and their backing components and use the same
ThreadPoolExecutor approach that we use for the component checkouts.

We aggressively parallelize the alias checkouts for aliases which use a
dependency URL that we've already cloned during the component checkout
phase, and group aliases with the same not-yet-cloned dependency URL to
avoid race conditions.

We rename the wrapper function which does the parallel work to
`do_parallel()` since it doesn't necessarily fetch anything anymore.

Additionally we add type annotations to `do_parallel()` to make it more
obvious how the arguments are supposed to look like.