Don't try to infer the dependencies directory in `component compile`
We mount the component repo at `/<component-name>` in the Docker
container when using `component compile` in the component template. In
that case, the old code inferred `dependencies_dir` as `/` which isn't
writable in the container environment.

This commit adjusts the implementation to use
`config.inventory.dependencies_dir` which should always be writable.
Since we don't actually use any functionality backed by
`MultiDependency` in `component compile` it shouldn't matter whether we
actually find the correct dependencies directory in all cases.