Update Dockerfile to use `commodore tool` to install external tools
We update the Dockerfile to use `commodore tool install` to install the
required external tools. Note that we use individual calls for each tool
so we can still have well-defined external tool versions in the
container image. To standardize the external tool management in the
container image, we also introduce an explicit Helm 3 version for the
container image.

Since we don't need Helm 2 anymore, we don't install the `helm2` binary
in the container image anymore.

Finally, we also remove `tools/install-jb.sh` which isn't required
anymore and adjust the container image test to use the new `commodore
version` command.