Fix permissions for `/app/.cache/` in the container image
We now populate `/app/.cache/commodore/tools` with the required external
tools, but other tooling (e.g. Helm) also needs to be able to write to
`/app/.cache` (and commodore itself needs to be able to write to
`/app/.cache/commodore/token`). While refactoring the Dockerfile, we
forgot to extend the permission adjustments for `/app` to also cover
existing directories in `/app/.cache`.

This commit ensures that everything in `/app` has group 0 and that
directories `/app/.cache/` and `/app/.cache/commodore` are writable by
group 0.